+++ changelog.md
@@ -48,6 +48,7 @@
 * Piglin Quartz -> Break Anger
   * Can now be fully configured with which blocks are broken and which mobs are angered
   * By default, it's the same as before: Zombified Piglins get angry if players break quartz
+* "Allow Climbing" is much more reliable and mobs no longer get stuckat the top of a ladder
 * Mobs are now affected by blindness (They see only 10% of range)
 * Drowned should get stuck less
 * Witches no longer heal themselves if above 70% health